+++ product/src/main/java/org/mobile/product/service/ProductService.java
@@ -3,16 +3,14 @@ package org.mobile.product.service;
 import java.util.List;
 
 import org.mobile.product.model.Product;
-
 public interface ProductService {
 	
 	List<Product> findAll();
 	
-	String delete(int id);
-	
+	String delete(Long id);
 	String save(Product product);
 	
-	Product findOne(int id);
+	Product findOne(Long id);
 
 	List<Product> findByCategoryId(int category_id);
